% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster.R
\name{clusterGenesetsByGenes}
\alias{clusterGenesetsByGenes}
\title{Cluster a set of genesets by gene membership}
\usage{
clusterGenesetsByGenes(results_table, genes_col = "gene_names",
  dist_method = "manhattan", hclust_method = "ward.D2")
}
\arguments{
\item{results_table}{The table of geneset results}

\item{genes_col}{The name of the column containing comma-separated genenames}

\item{dist_method}{The name of the distance method to pass to "dist"}

\item{hclust_method}{The name of the clustering method to pass to "hclust"}
}
\description{
Cluster a set of genesets by gene membership
}
