% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{fetchAnnotation}
\alias{fetchAnnotation}
\title{Fetch table of ensembl_ids, entrez_ids and gene_names using bioMart.}
\usage{
fetchAnnotation(species = "hs", ensembl_version = "latest")
}
\arguments{
\item{species}{The species code, only "hs" or "mm" are supported.}

\item{ensembl_version}{The version of ENSEMBL annotation to use.}
}
\description{
Fetch table of ensembl_ids, entrez_ids and gene_names using bioMart.
}
\examples{
annotation <- fetchAnnotation(species="hs")
}
