% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations2.R
\name{fetchKeggGeneSets}
\alias{fetchKeggGeneSets}
\title{Fetch KEGG annotations into list of gene sets}
\usage{
fetchKeggGeneSets(species = c("hs", "mm"))
}
\arguments{
\item{species}{Species identifier (only "hs" or "mm" are supported).}
}
\value{
A named list of gene sets.
}
\description{
Fetch KEGG annotations into list of gene sets
}
\examples{
\dontrun{
kegg_hs <- fetchKeggGeneSets(species="hs")
kegg_mm <- fetchKeggGeneSets(species="mm")
}
}
