% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\name{filterGenesets}
\alias{filterGenesets}
\title{Filter a genesets results table}
\usage{
filterGenesets(results_table, min_foreground_genes = 3,
  max_genes_geneset = 500, min_odds_ratio = 2, p_col = "p.val",
  adjust_pvalues = FALSE, padjust_method = "BH",
  pvalue_threshold = 0.05)
}
\arguments{
\item{results_table}{A \emph{gofisher} multi-sample results table.}

\item{min_foreground_genes}{Minimum number of over-represent genes in foreground.}

\item{max_genes_geneset}{Maximum number of genes in the full geneset (used to remove generic categories).}

\item{min_odds_ratio}{Only genesets with odds ratios equal to or greater will be considered.}

\item{adjust_pvalues}{Logical value indicating whether to adjust p-values
across all sample.}

\item{padjust_method}{Correction method. Can be abbreviated
(see \code{\link{p.adjust.methods}}).}

\item{pvalue_threshold}{P-value threshold.}
}
\description{
Filter a genesets results table
}
\seealso{

}
\author{
Steve Sansom
}
