% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{formatDescriptions}
\alias{formatDescriptions}
\title{Utility function to clean up desriptions for plots}
\usage{
formatDescriptions(xx, remove = c(), maxl = 45)
}
\arguments{
\item{remove}{Vector of string patterns to delete}

\item{maxl}{Max length of the descriptions}

\item{descriptions}{Vector of descriptions to clean}
}
\description{
Utility function to clean up desriptions for plots
}
