% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{getENSEMBL}
\alias{getENSEMBL}
\title{Get the org.Xx.eg.db ENSEMBL bimap object}
\usage{
getENSEMBL(species = c("mm", "hs"))
}
\arguments{
\item{species}{Species, "mm" or "hs"}
}
\description{
Get the org.Xx.eg.db ENSEMBL bimap object
}
