% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{getEntrez}
\alias{getEntrez}
\title{Verify input gene id type and return entrez identifiers}
\usage{
getEntrez(
  gene_ids,
  gene_id_type = c("ensembl", "entrez"),
  species = c("mm", "hs")
)
}
\arguments{
\item{gene_ids}{A character vector of gene identifiers}

\item{gene_id_type}{Either "ensembl" or "entrez"}

\item{species}{Either "mm" or "hs"}
}
\description{
Verify input gene id type and return entrez identifiers
}
