% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{getGO}
\alias{getGO}
\title{Get the org.Xx.eg.db GO2ALLEGS bimap object.}
\usage{
getGO(species = c("mm", "hs"))
}
\arguments{
\item{species}{Species, "mm" or "hs"}
}
\description{
Get the org.Xx.eg.db GO2ALLEGS bimap object.
}
