% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{getSampleGenesets}
\alias{getSampleGenesets}
\title{Get a list of top-genesets by cluster}
\usage{
getSampleGenesets(
  results_table,
  sort_by = c("p.val", "p.adj", "odds.ratio", "score"),
  sample_id_col = "cluster",
  max_rows = 50
)
}
\arguments{
\item{sort_by}{The column by which to sort the genesets}
}
\description{
Get a list of top-genesets by cluster
}
