% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations2.R
\name{makeKeggCollectionFromList}
\alias{makeKeggCollectionFromList}
\title{Make a KEGGCollection from a list of gene sets}
\usage{
makeKeggCollectionFromList(x)
}
\arguments{
\item{x}{A list of KEGG pathway gene sets
(e.g. from \code{fetchKeggGeneSets}).}
}
\value{
A \code{\link{KEGGCollection}}
}
\description{
Make a KEGGCollection from a list of gene sets
}
\note{
This function is hard-coded to use Entrez gene identifiers,
which is expected by \emph{gsfisher}.
It won't fail if that is not the case, but this may impact downstream
analyses.
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- read.gmt(gmtFile)

kc <- makeKeggCollectionFromList(ann_gmt)
}
