% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{mapENTREZhuman2mouse}
\alias{mapENTREZhuman2mouse}
\title{Translate ENTREZ gene sets from human to mouse}
\usage{
mapENTREZhuman2mouse(GMT, ensembl_version = NULL)
}
\arguments{
\item{GMT}{A named list of gene sets
(e.g., from \code{readGMT}).}

\item{ensembl_version}{Version of the ensembl annotation to use,
passed to \code{biomaRt::useEnsembl}.
The default \code{NULL} uses the current annotation release.}
}
\description{
Uses ENSEMBL biomaRt for converting human MSigDB gene sets to mouse.
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- readGMT(gmtFile)
\dontrun{
mapENTREZhuman2mouse(ann_gmt, ensembl_version=NULL)
}
}
\seealso{
\code{\link{readGMT}}
\code{\link[biomaRt]{useEnsembl}}
}
\author{
Steve Sansom
}
