% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{read.gmt}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{qusage}{\code{\link[qusage]{read.gmt}}}
}}

