% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{runGMT.all}
\alias{runGMT.all}
\title{Run GMT analysis on a multi-sample results table.}
\usage{
runGMT.all(results = NULL, species = c("mm", "hs"), background_ids = NULL,
  sample_col = "cluster", gmt_file = NULL, p_col = "p_val_adj",
  p_threshold = 0.1)
}
\arguments{
\item{results}{A multi-sample results table}

\item{species}{Either "mm" or "hs"}

\item{background_ids}{A vector of ENSEMBL gene ids. If NULL, taken from results.}

\item{sample_col}{The column in results that indicates the sample}

\item{gmt_file}{A GMT file.}

\item{p_col}{The column containing the p-values to use}

\item{p_threshold}{The significance threshold.}
}
\description{
Run GMT analysis on a multi-sample results table.
}
