% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{runGMT}
\alias{runGMT}
\title{A wrapper function to run Fisher tests for enrichement from GMT files}
\usage{
runGMT(foreground_ids, background_ids, gmt_file, annotation)
}
\arguments{
\item{foreground_ids}{The list of entrez ids of interest (e.g. significantly differentially expressed genes).}

\item{background_ids}{The list of entrez ids againt which enrichment will be tested (i.e. the gene universe).}

\item{gmt_file}{The location of the GMT file.}

\item{annotation}{A dataframe with columns "entrez_id" and "gene_name" (see fetchAnnotation).}
}
\description{
A wrapper function to run Fisher tests for enrichement from GMT files
}
