% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{runGO.all}
\alias{runGO.all}
\title{Run GO analysis on a multi-sample results table.}
\usage{
runGO.all(results = NULL, species = c("mm", "hs"), background_ids = NULL,
  gene_id_col = "gene_id", gene_id_type = c("entrez", "ensembl"),
  sample_col = "cluster", p_col = "p_val_adj", p_threshold = 0.1)
}
\arguments{
\item{results}{A multi-sample results table}

\item{species}{Either "mm" or "hs"}

\item{background_ids}{A vector of gene ids. If NULL, taken from results.}

\item{gene_id_col}{The name of the column containing the gene identifiers.}

\item{gene_id_type}{Either "entrez" (default) or "ensembl".}

\item{sample_col}{The column in results that indicates the sample}

\item{p_col}{The column containing the p-values to use}

\item{p_threshold}{The significance threshold.}
}
\description{
Run GO analysis on a multi-sample results table.
}
