% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO.R
\name{runGO}
\alias{runGO}
\title{Run gene set enrichement on Gene Ontology categories}
\usage{
runGO(foreground_ids, background_ids, gene_id_type = c("entrez", "ensembl"),
  species = c("hs", "mm"), ...)
}
\arguments{
\item{foreground_ids}{A list of ENTREZ ids of interest
(e.g. significantly differentially expressed genes).}

\item{background_ids}{A list of background ENTREZ ids.
against which enrichment will be tested (i.e., the gene universe).
(see \code{fetchAnnotation}).}

\item{gene_id_type}{Either "entrez" (default") or "ensembl".}

\item{species}{Species identifier (only "hs" or "mm" are supported).}
}
\description{
A wrapper function to run Fisher's test for enrichement
on Gene Ontology (GO) categories.
Depends on the bioconductor \code{org.xx.eg.db} and \code{GO.db} packages.
}
\examples{
# TODO
\dontrun{
# TODO
}
}
\seealso{
\code{\link{runFisherTests}}.
}
\author{
Steve Sansom
}
