% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{runKEGG.all}
\alias{runKEGG.all}
\title{Run KEGG analysis on a multi-sample results table.}
\usage{
runKEGG.all(results = NULL, species = c("mm", "hs"),
  background_ids = NULL, sample_col = "cluster",
  gene_id_col = "gene_id", gene_id_type = c("entrez", "ensembl"),
  keggData = NULL, p_col = "p_val_adj", p_threshold = 0.1)
}
\arguments{
\item{results}{A multi-sample results table}

\item{species}{Either "mm" or "hs"}

\item{background_ids}{A vector of ENSEMBL gene ids. If NULL, taken from results.}

\item{sample_col}{The column in results that indicates the sample}

\item{gene_id_col}{The column containing the gene identifiers}

\item{gene_id_type}{Either "entrez" (default) or "ensembl".}

\item{keggData}{A list of KEGG gene sets.
(see \code{fetchKegg}).}

\item{p_col}{The column containing the p-values to use}

\item{p_threshold}{The significance threshold.}
}
\description{
Run KEGG analysis on a multi-sample results table.
}
