% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{sampleEnrichmentDotplot}
\alias{sampleEnrichmentDotplot}
\title{Make a dot-plot of geneset enrichments across multiple samples/clusters}
\usage{
sampleEnrichmentDotplot(results_table, p_col = "p.adj",
  pvalue_threshold = 0.05, selected_genesets = c(),
  sample_id_col = "cluster")
}
\arguments{
\item{results_table}{The results table}

\item{p_col}{The name of the column with the (e.g. adjusted) p-values}

\item{pvalue_threshold}{The pvalue_threshold to be applied}

\item{selected_genesets}{A vector of geneset descriptions to slow in the plot}

\item{sample_id_col}{The column of the results_table containing the sample id}
}
\description{
Make a dot-plot of geneset enrichments across multiple samples/clusters
}
