% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{sampleEnrichmentHeatmap}
\alias{sampleEnrichmentHeatmap}
\title{Draw a heatmap of enriched genesets in all samples}
\usage{
sampleEnrichmentHeatmap(results_table, max_rows = 50, min_genes = 2,
  adjust_pvalues = FALSE, padjust_method = "BH", pvalue_threshold = 0.1,
  maxl = 45, show_common = TRUE, sample_id_col = "cluster",
  title = "enriched genesets")
}
\arguments{
\item{results_table}{A gofisher multi-sample results table}

\item{max_rows}{The maximum number of rows to show in the heatmap}

\item{adjust_pvalues}{Adjust p-values across all sample}

\item{pvalue_threshold}{P-value threshold}

\item{maxl}{The maximum length for geneset names}

\item{show_common}{Show genesets enriched in all samples.}

\item{sample_id_col}{The column of the results_table containing the sample id}
}
\description{
Draw a heatmap of enriched genesets in all samples
}
