% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{translateGMT2mouse}
\alias{translateGMT2mouse}
\title{Translate a list of genesets from human to mouse entrezgenes.}
\usage{
translateGMT2mouse(GMT, ensembl_version = "latest")
}
\arguments{
\item{GMT}{A GMT-like geneset list (e.g. from readGMT)}

\item{ensembl_version}{Version of the ensembl annotation to use.}
}
\description{
Useful for converting human msigdb GMTs to mouse.
}
