% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{visualiseClusteredGenesets}
\alias{visualiseClusteredGenesets}
\title{Visulalise a set of clustered genesets on a circular dendrogram}
\usage{
visualiseClusteredGenesets(results_table, geneset_clust = NULL,
  odds_ratio_max_quantile = 0.9, highlight = c())
}
\arguments{
\item{geneset_clust}{A hclust object (e.g. from "clusterGenesetsByGenes").}

\item{odds_ratio_max_quantile}{A quantile to cap the odds.ratio at for visulisation purposes.}

\item{highlight}{A vector of descriptions to highlight in the plot}
}
\description{
Visulalise a set of clustered genesets on a circular dendrogram
}
