% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotations.R
\name{fetchAnnotation}
\alias{fetchAnnotation}
\title{Fetch ENSEMBL ids, ENTREZ ids and gene names using bioMart}
\usage{
fetchAnnotation(species = c("hs", "mm"), ensembl_version = NULL,
  ensembl_host = NULL)
}
\arguments{
\item{species}{Species identifier (only "hs" or "mm" are supported).}

\item{ensembl_version}{Version of the ensembl annotation to use,
passed to \code{biomaRt::useEnsembl}.
The default \code{NULL} uses the current annotation release.}

\item{ensembl_host}{The address of the ensembl host
The default \code{NULL} uses www.ensembl.org}
}
\value{
A data.frame of three columns:
\code{c("ensembl_id", "entrez_id", "gene_name")}
}
\description{
Fetches annotations for human ("hs") or mouse ("mm") from the
Ensembl BioMart.
}
\details{
Only human "hs" and mouse "mm" are supported.
Annotation must contain the columns \code{"entrez_id"} and
\code{"gene_name"}.
}
\examples{
\dontrun{
ann_hs <- fetchAnnotation(species="hs")
ann_mm <- fetchAnnotation(species="mm")
}
}
\author{
Steve Sansom
}
