% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{fetchAnnotation}
\alias{fetchAnnotation}
\title{Fetch table of ensembl_ids, entrez_ids and gene_names using bioMart.}
\usage{
fetchAnnotation(species = c("hs", "mm"), ensembl_version = "latest")
}
\arguments{
\item{species}{The species code, only "hs" or "mm" are supported.}

\item{ensembl_version}{The version of ENSEMBL annotation to use.}
}
\description{
Fetch table of ensembl_ids, entrez_ids and gene_names using bioMart.
}
\details{
Only human "hs" and mouse "mm" are supported.
Annotation must contain the columns \code{"entrez_id"} and
\code{"gene_name"}.
}
\examples{
\dontrun{
ann_hs <- fetchAnnotation(species="hs")
ann_mm <- fetchAnnotation(species="mm")}
}
