% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{fisherTest}
\alias{fisherTest}
\title{Perform a single Fisher tests for geneset enrichement.}
\usage{
fisherTest(n, genesets = "none", foreground_ids = "none",
  background_ids = "none", annotation = "none")
}
\arguments{
\item{n}{the index of the geneset to test}

\item{genesets}{the list of genesets}

\item{foreground_ids}{the list of entrez ids of interest (e.g. significantly differentially expressed genes)}

\item{background_ids}{the list of entrez ids againt which enrichment will be tested (i.e. the gene universe)}

\item{annotation}{a dataframe with columns "entrez_id" and "gene_name" (see fetchAnnotation)}
}
\description{
Perform a single Fisher tests for geneset enrichement.
}
