% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{readGMT}
\alias{readGMT}
\title{Read genesets from a GMT file}
\usage{
readGMT(filepath)
}
\arguments{
\item{filepath}{The location of the GMT file.}
}
\description{
Read genesets from a GMT file
}
\details{
Deprecated. Use \code{qusage::read.gmt} instead.
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- readGMT(gmtFile)
}
\seealso{
\code{\link[qusage]{read.gmt}}
}
