% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{runFisherTests}
\alias{runFisherTests}
\title{Run a set of Fisher tests for gene set enrichement}
\usage{
runFisherTests(named_geneset_list, foreground_ids, background_ids, SYMBOL)
}
\arguments{
\item{named_geneset_list}{List of named gene sets.}

\item{foreground_ids}{A list of ENTREZ ids of interest
(e.g. significantly differentially expressed genes)}

\item{background_ids}{A list of background ENTREZ ids.
against which enrichment will be tested (i.e., the gene universe)}

\item{annotations}{A data.frame with at least columns "entrez_id" and "gene_name"
(see \code{fetchAnnotation}).}
}
\description{
Run a set of Fisher tests for gene set enrichement
}
\examples{
# TODO
\dontrun{
# TODO
}
}
\seealso{
\code{\link{readGMT}},
\code{\link{fisherTest}}.
}
\author{
Steve Sansom
}
