% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{runFisherTests}
\alias{runFisherTests}
\title{Run a set of Fisher tests for gene set enrichement}
\usage{
runFisherTests(named_geneset_list, foreground_ids, background_ids, annotations)
}
\arguments{
\item{named_geneset_list}{List of named gene sets.}

\item{foreground_ids}{A list of ENTREZ ids of interest
(e.g. significantly differentially expressed genes)}

\item{background_ids}{A list of background ENTREZ ids.
against which enrichment will be tested (i.e., the gene universe)}

\item{annotations}{A data.frame with at least columns "entrez_id" and "gene_name"
(see \code{fetchAnnotations}).}
}
\description{
Run a set of Fisher tests for gene set enrichement
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- readGMT(gmtFile)

# Take 50\% of the first gene set as an example list of interest
foreground <- head(ann_gmt[[1]], length(ann_gmt[[1]]) / 2)

\dontrun{
# Fetch annotations
ann_hs <- fetchAnnotations(species="hs")
background <- subset(ann_hs, !is.na(entrez_id), "entrez_id", drop=TRUE)
result <- runFisherTests(ann_gmt, foreground, background, ann_hs)
}
}
