% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{runFisherTests}
\alias{runFisherTests}
\title{Run a set of Fisher tests for geneset enrichement.}
\usage{
runFisherTests(named_geneset_list, foreground_ids, background_ids, annotation)
}
\arguments{
\item{named_geneset_list}{List of named gene sets.}

\item{foreground_ids}{the list of entrez ids of interest (e.g. significantly differentially expressed genes)}

\item{background_ids}{the list of entrez ids againt which enrichment will be tested (i.e. the gene universe)}

\item{annotation}{a dataframe with columns "entrez_id" and "gene_name" (see fetchAnnotation)}
}
\description{
Run a set of Fisher tests for geneset enrichement.
}
