% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{runGMT}
\alias{runGMT}
\title{Run gene set enrichement on custom gene sets}
\usage{
runGMT(foreground_ids, background_ids, gmt_file, annotations)
}
\arguments{
\item{foreground_ids}{A list of ENTREZ ids of interest
(e.g. significantly differentially expressed genes).}

\item{background_ids}{A list of background ENTREZ ids.
against which enrichment will be tested (i.e., the gene universe).}

\item{gmt_file}{path to a GMT file.}

\item{annotations}{A data.frame with at least columns "entrez_id" and "gene_name"
(see \code{fetchAnnotations}).}
}
\description{
A wrapper function to run Fisher tests for enrichement from GMT files.
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- readGMT(gmtFile)

# Take 50\% of the first gene set as an example list of interest
foreground <- head(ann_gmt[[1]], length(ann_gmt[[1]]) / 2)

\dontrun{
# Fetch annotations
ann_hs <- fetchAnnotations(species="hs")
background <- subset(ann_hs, !is.na(entrez_id), "entrez_id", drop=TRUE)
result <- runGMT(foreground, background, gmtFile, ann_hs)
}
}
\author{
Steve Sansom and Kevin-Rue-Albrecht
}
