% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{runGO}
\alias{runGO}
\title{A wrapper function to run Fisher tests for enrichement of Gene Ontology (GO) categories.
Depends on the bioconductor org.xx.eg.db and GO.db libraries}
\usage{
runGO(foreground_ids, background_ids, annotation, species = c("hs", "mm"))
}
\arguments{
\item{foreground_ids}{the list of entrez ids of interest (e.g. significantly differentially expressed genes)}

\item{background_ids}{the list of entrez ids againt which enrichment will be tested (i.e. the gene universe)}

\item{annotation}{a dataframe with columns "entrez_id" and "gene_name" (see fetchAnnotations)}

\item{species}{The species code, only "hs" or "mm" are supported.}
}
\description{
A wrapper function to run Fisher tests for enrichement of Gene Ontology (GO) categories.
Depends on the bioconductor org.xx.eg.db and GO.db libraries
}
