% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{runGO}
\alias{runGO}
\title{Run Gene Ontology gene set enrichement}
\usage{
runGO(foreground_ids, background_ids, annotations, species = c("hs", "mm"))
}
\arguments{
\item{foreground_ids}{A list of ENTREZ ids of interest
(e.g. significantly differentially expressed genes)}

\item{background_ids}{A list of background ENTREZ ids.
against which enrichment will be tested (i.e., the gene universe)}

\item{annotations}{A data.frame with at least columns "entrez_id" and "gene_name"
(see \code{fetchAnnotations}).}

\item{species}{Species identifiers (only "hs" or "mm" are supported).}
}
\description{
A wrapper function to run Fisher's test for enrichement
on Gene Ontology (GO) categories.
Depends on the bioconductor \code{org.xx.eg.db} and \code{GO.db} packages.
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- readGMT(gmtFile)

index <- 1
foreground <- head(ann_gmt[[example]], length(ann_gmt[[example]]) / 2)

\dontrun{
# Fetch annotations
ann_hs <- fetchAnnotations(species="hs")
background <- subset(ann_hs, !is.na(entrez_id), "entrez_id", drop=TRUE)
runGO(foreground, background, ann_hs, "hs")
}
}
