% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG.R
\name{runKEGG}
\alias{runKEGG}
\title{Run gene set enrichement on KEGG pathways}
\usage{
runKEGG(foreground_ids, background_ids, keggData = NULL, SYMBOL = NULL,
  gene_id_type = c("entrez", "ensembl"), species = c("hs", "mm"))
}
\arguments{
\item{foreground_ids}{A list of ENTREZ ids of interest
(e.g. significantly differentially expressed genes).}

\item{background_ids}{A list of background ENTREZ ids.
against which enrichment will be tested (i.e., the gene universe).}

\item{keggData}{A list of KEGG gene sets.}

\item{gene_id_type}{Either "entrez" (default) or "ensembl".}

\item{species}{Species identifier (only "hs" or "mm" are supported).}
}
\description{
A wrapper function to run Fisher tests for enrichement of KEGG Pathways.
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- readGMT(gmtFile)
# Take 50\% of the first gene set as an example list of interest
foreground <- head(ann_gmt[[1]], length(ann_gmt[[1]]) / 2)
\dontrun{
# TODO
}
}
\seealso{
\code{\link{readGMT}},
\code{\link{fetchKEGG}},
\code{\link{runFisherTests}}.
}
\author{
Steve Sansom
}
