% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{sampleEnrichmentDotplot}
\alias{sampleEnrichmentDotplot}
\title{Make a dot-plot of geneset enrichments across multiple samples/clusters}
\usage{
sampleEnrichmentDotplot(results_table, p_col = "p.adj",
  pvalue_threshold = 0.05, selected_genesets = c(),
  selection_col = "description", sample_id_col = "cluster",
  sample_levels = NULL, fill_var = "odds.ratio", fill_trans = "log2",
  fill_colors = c("yellow", "red", "brown"), fill_max_quantile = 1,
  non_significant_color = "grey90", min_dot_size = 2,
  max_dot_size = 8, maxl = Inf, title = NULL)
}
\arguments{
\item{results_table}{The results table}

\item{p_col}{The name of the column with the (e.g. adjusted) p-values}

\item{pvalue_threshold}{The pvalue_threshold to be applied}

\item{selected_genesets}{A vector of geneset slow in the plot}

\item{selection_col}{The column against which the geneset selections will be matched}

\item{sample_id_col}{The column of the results_table containing the sample id}

\item{fill_var}{The column to use for the fill color.}

\item{fill_trans}{Name of the transformation for the fill. "-log10" is supported for p-values.}

\item{fill_colors}{Colors for the fill gradient}

\item{fill_max_quantile}{The quantile at which (non-infinite) fill variable will be capped for the color scale.}

\item{min_dot_size}{Minimum dot size (points)}

\item{max_dot_size}{Maximum dot size (points)}

\item{maxl}{The maximum lenght for the geneset labels}
}
\description{
Make a dot-plot of geneset enrichments across multiple samples/clusters
}
