% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{sampleEnrichmentHeatmap}
\alias{sampleEnrichmentHeatmap}
\title{Draw a heatmap of enriched genesets in all samples}
\usage{
sampleEnrichmentHeatmap(results_table, max_rows = 50, min_genes = 2,
  p_col = "p.val", adjust_pvalues = FALSE, padjust_method = "BH",
  pvalue_threshold = 0.1, maxl = 45, show_common = TRUE,
  sample_id_col = "cluster", title = "Enriched gene sets")
}
\arguments{
\item{results_table}{A \emph{gofisher} multi-sample results table.}

\item{max_rows}{Maximum number of rows to show in the heatmap.}

\item{min_genes}{Minimum number of genes to filter gene sets.}

\item{adjust_pvalues}{Logical value indicating whether to adjust p-values
across all sample.}

\item{padjust_method}{Correction method. Can be abbreviated
(see \code{\link{p.adjust.methods}}).}

\item{pvalue_threshold}{P-value threshold.}

\item{maxl}{Maximum length for gene set names (trimmed otherwise).}

\item{show_common}{Logical value indicating whether to show gene sets
enriched in all samples.}

\item{sample_id_col}{The column of the results_table containing the sample id}

\item{title}{Title of the heat map plot.}
}
\description{
Draw a heatmap of enriched genesets in all samples
}
\seealso{
\code{\link{readGMT}}
\code{\link{p.adjust}}
}
\author{
Steve Sansom
}
