% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{translateGMT2mouse}
\alias{translateGMT2mouse}
\title{Translate a list of genesets from human to mouse entrezgenes.}
\usage{
translateGMT2mouse(GMT, ensembl_version = "latest")
}
\arguments{
\item{GMT}{A GMT-like gene set list (e.g., from \code{readGMT})}

\item{ensembl_version}{Version of the ensembl annotation to use.}
}
\description{
Useful for converting human msigdb GMTs to mouse.
}
\note{
Use \code{qusage::read.gmt} instead of \code{readGMT}.
}
