% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genesetFisherTests.R
\name{translateGMT2mouse}
\alias{translateGMT2mouse}
\title{Translate a list of genesets from human to mouse entrezgenes.}
\usage{
translateGMT2mouse(GMT, ensembl_version = "latest")
}
\arguments{
\item{GMT}{A named list of gene sets
(e.g., from \code{qusage::read.gmt}).}

\item{ensembl_version}{Version of the ensembl annotation to use
(see \code{biomaRt::useEnsembl}).}
}
\description{
Useful for converting human msigdb GMTs to mouse.
}
\seealso{
\code{\link[qusage]{read.gmt}}
\code{\link[biomaRt]{useEnsembl}}
}
