% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{visualiseClusteredGenesets}
\alias{visualiseClusteredGenesets}
\title{Visulalise a set of clustered genesets on a circular dendrogram}
\usage{
visualiseClusteredGenesets(results_table, geneset_clust = NULL,
  odds_ratio_max_quantile = 0.9, highlight = c(),
  id_col = "geneset_id", name_col = "description",
  ngenes_col = "n_fg", odds_ratio_col = "odds.ratio")
}
\arguments{
\item{results_table}{A table of geneset results.}

\item{geneset_clust}{A hclust object (e.g. from "clusterGenesetsByGenes").}

\item{odds_ratio_max_quantile}{A quantile to cap the odds.ratio at for visulisation purposes.}

\item{highlight}{A vector of descriptions to highlight in the plot}

\item{id_col}{The column containing the unique geneset identifiers}

\item{name_col}{The column containing the geneset name}

\item{ngenes_col}{The column containing the number of geneset genes in the foreground}

\item{odds_ratio_col}{The column containing the odds ratio for geneset enrichment}
}
\description{
Visulalise a set of clustered genesets on a circular dendrogram
}
