% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{writeGMT}
\alias{writeGMT}
\title{Write gene sets to a GMT file}
\usage{
writeGMT(geneset, outfile)
}
\arguments{
\item{geneset}{A list of gene sets  (e.g. from \code{qusage::read.gmt}).}

\item{outfile}{Path to the output file.}
}
\description{
Write gene sets to a GMT file
}
\examples{
gmtFile <- system.file(package = "gsfisher", "extdata", "kegg_hs.gmt")
ann_gmt <- read.gmt(gmtFile)

outfile <- tempfile()
writeGMT(ann_gmt, outfile)
}
\seealso{
\code{\link{read.gmt}}
}
\author{
Steve Sansom
}
