% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMT.R
\name{writeGMT}
\alias{writeGMT}
\title{Write genesets to a GMT file}
\usage{
writeGMT(geneset, outfile)
}
\arguments{
\item{geneset}{A geneset list (e.g. from readGMT)}

\item{outfile}{The name of the outfile}
}
\description{
Write genesets to a GMT file
}
